feat: show different settings for different modes
