+++ WebUI/src/assets/i18n/de.json
@@ -251,5 +251,8 @@
   "DEVICE": "Gerät",
   "SETTINGS_IMAGE_COMFY_CHECKPOINT_MODEL": "Checkpoint-Modell",
   "SETTINGS_IMAGE_COMFY_FRAME_RATE": "Bildrate",
-  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "Gesamtanzahl Frames"
+  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "Gesamtanzahl Frames",
+  "SETTINGS_LLM_MAX_TOKENS": "Max. Tokens",
+  "SETTINGS_LLM_PERFORMANCE_METRICS": "Performance-Metriken",
+  "SETTINGS_TAB_LLM": "LLM"
 }
