+++ WebUI/src/assets/i18n/en-US.json
@@ -251,5 +251,8 @@
   "DEVICE": "Device",
   "SETTINGS_IMAGE_COMFY_CHECKPOINT_MODEL": "Checkpoint Model",
   "SETTINGS_IMAGE_COMFY_FRAME_RATE": "Frame Rate",
-  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "Total Frames"
+  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "Total Frames",
+  "SETTINGS_LLM_MAX_TOKENS": "Max Tokens",
+  "SETTINGS_LLM_PERFORMANCE_METRICS": "Performance Metrics",
+  "SETTINGS_TAB_LLM": "LLM"
 }
