+++ WebUI/src/assets/i18n/es.json
@@ -251,5 +251,8 @@
   "DEVICE": "Dispositivo",
   "SETTINGS_IMAGE_COMFY_CHECKPOINT_MODEL": "Modelo de checkpoint",
   "SETTINGS_IMAGE_COMFY_FRAME_RATE": "Velocidad de fotogramas",
-  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "Frames totales"
+  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "Frames totales",
+  "SETTINGS_LLM_MAX_TOKENS": "Tokens máximos",
+  "SETTINGS_LLM_PERFORMANCE_METRICS": "Métricas de rendimiento",
+  "SETTINGS_TAB_LLM": "LLM"
 }
