+++ WebUI/src/assets/i18n/id.json
@@ -251,5 +251,8 @@
   "DEVICE": "Perangkat",
   "SETTINGS_IMAGE_COMFY_CHECKPOINT_MODEL": "Model Checkpoint",
   "SETTINGS_IMAGE_COMFY_FRAME_RATE": "Frame Rate",
-  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "Total Frame"
+  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "Total Frame",
+  "SETTINGS_LLM_MAX_TOKENS": "Token Maksimum",
+  "SETTINGS_LLM_PERFORMANCE_METRICS": "Metrik Performa",
+  "SETTINGS_TAB_LLM": "LLM"
 }
