+++ WebUI/src/assets/i18n/ja.json
@@ -251,5 +251,8 @@
   "DEVICE": "デバイス",
   "SETTINGS_IMAGE_COMFY_CHECKPOINT_MODEL": "チェックポイントモデル",
   "SETTINGS_IMAGE_COMFY_FRAME_RATE": "フレームレート",
-  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "総フレーム数"
+  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "総フレーム数",
+  "SETTINGS_LLM_MAX_TOKENS": "最大トークン数",
+  "SETTINGS_LLM_PERFORMANCE_METRICS": "パフォーマンス指標",
+  "SETTINGS_TAB_LLM": "LLM"
 }
