+++ WebUI/src/assets/i18n/ko.json
@@ -251,5 +251,8 @@
   "DEVICE": "디바이스",
   "SETTINGS_IMAGE_COMFY_CHECKPOINT_MODEL": "체크포인트 모델",
   "SETTINGS_IMAGE_COMFY_FRAME_RATE": "프레임 레이트",
-  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "총 프레임"
+  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "총 프레임",
+  "SETTINGS_LLM_MAX_TOKENS": "최대 토큰",
+  "SETTINGS_LLM_PERFORMANCE_METRICS": "성능 지표",
+  "SETTINGS_TAB_LLM": "LLM"
 }
