+++ WebUI/src/assets/i18n/tr.json
@@ -251,5 +251,8 @@
   "DEVICE": "Cihaz",
   "SETTINGS_IMAGE_COMFY_CHECKPOINT_MODEL": "Kontrol Noktası Modeli",
   "SETTINGS_IMAGE_COMFY_FRAME_RATE": "Kare Hızı",
-  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "Toplam Kare"
+  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "Toplam Kare",
+  "SETTINGS_LLM_MAX_TOKENS": "Maksimum Token",
+  "SETTINGS_LLM_PERFORMANCE_METRICS": "Performans Metrikleri",
+  "SETTINGS_TAB_LLM": "LLM"
 }
