+++ WebUI/src/assets/i18n/vi.json
@@ -251,5 +251,8 @@
   "DEVICE": "Thiết bị",
   "SETTINGS_IMAGE_COMFY_CHECKPOINT_MODEL": "Mô hình checkpoint",
   "SETTINGS_IMAGE_COMFY_FRAME_RATE": "Tốc độ khung hình",
-  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "Tổng số khung hình"
+  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "Tổng số khung hình",
+  "SETTINGS_LLM_MAX_TOKENS": "Token tối đa",
+  "SETTINGS_LLM_PERFORMANCE_METRICS": "Chỉ số hiệu suất",
+  "SETTINGS_TAB_LLM": "LLM"
 }
