+++ WebUI/src/assets/i18n/zh-CN.json
@@ -251,5 +251,8 @@
   "DEVICE": "设备",
   "SETTINGS_IMAGE_COMFY_CHECKPOINT_MODEL": "检查点模型",
   "SETTINGS_IMAGE_COMFY_FRAME_RATE": "帧率",
-  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "总帧数"
+  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "总帧数",
+  "SETTINGS_LLM_MAX_TOKENS": "最大标记数",
+  "SETTINGS_LLM_PERFORMANCE_METRICS": "性能指标",
+  "SETTINGS_TAB_LLM": "LLM"
 }
