+++ WebUI/src/assets/i18n/zh-TW.json
@@ -251,5 +251,8 @@
   "DEVICE": "設備",
   "SETTINGS_IMAGE_COMFY_CHECKPOINT_MODEL": "檢查點模型",
   "SETTINGS_IMAGE_COMFY_FRAME_RATE": "幀率",
-  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "總幀數"
+  "SETTINGS_IMAGE_COMFY_TOTAL_FRAMES": "總幀數",
+  "SETTINGS_LLM_MAX_TOKENS": "最大標記數",
+  "SETTINGS_LLM_PERFORMANCE_METRICS": "效能指標",
+  "SETTINGS_TAB_LLM": "LLM"
 }
