+++ WebUI/src/components/SettingsBasic.vue
@@ -40,102 +40,27 @@
   </div>
   <div class="border-b border-color-spilter flex flex-col gap-5 py-4">
     <h2 class="text-center font-bold">{{ languages.SETTINGS_BASIC_BACKEND }}</h2>
-    <div class="flex flex-col gap-2">
-      <p>{{ languages.SETTINGS_LLM_BACKEND }}</p>
-      <div class="flex items-center gap-2">
-        <drop-selector
-          :array="[...llmBackendTypes]"
-          @change="(item) => (textInference.backend = item)"
-        >
-          <template #selected>
-            <div class="flex gap-2 items-center">
-              <span
-                class="rounded-full w-2 h-2"
-                :class="{
-                  'bg-green-500': isRunning(textInference.backend),
-                  'bg-gray-500': !isRunning(textInference.backend),
-                }"
-              ></span>
-              <span>{{ textInferenceBackendDisplayName[textInference.backend] }}</span>
-              <!--       Flag LlamaCpp as experimental       -->
-              <span
-                v-if="textInference.backend == 'llamaCPP'"
-                class="rounded-lg h-4 px-1 text-xs"
-                :style="{ 'background-color': '#cc00ff88' }"
-              >
-                Experimental</span
-              >
-            </div>
-          </template>
-          <template #list="slotItem">
-            <div class="flex gap-2 items-center">
-              <span
-                class="rounded-full w-2 h-2"
-                :class="{
-                  'bg-green-500': isRunning(slotItem.item),
-                  'bg-gray-500': !isRunning(slotItem.item),
-                }"
-              ></span>
-              <span>{{
-                textInferenceBackendDisplayName[slotItem.item as (typeof llmBackendTypes)[number]]
-              }}</span>
-              <span
-                v-if="slotItem.item == 'llamaCPP'"
-                class="rounded-lg h-4 px-1 text-xs"
-                :style="{ 'background-color': '#cc00ff88' }"
-              >
-                Experimental</span
-              >
-            </div>
-          </template>
-        </drop-selector>
+    <div class="flex flex-col gap-3">
+      <p>{{ languages.SETTINGS_BACKEND_STATUS }}</p>
+      <table class="text-center w-full mx-2 table-fixed">
+        <tbody>
+          <tr v-for="item in displayComponents" :key="item.serviceName">
+            <td style="text-align: left">{{ mapServiceNameToDisplayName(item.serviceName) }}</td>
+            <td :style="{ color: mapStatusToColor(item.status) }">
+              {{ mapToDisplayStatus(item.status) }}
+            </td>
+          </tr>
+        </tbody>
+      </table>
+      <div class="flex flex-col pt-5">
+        <button @click="globalSetup.loadingState = 'manageInstallations'" class="confirm-btn">
+          {{ languages.SETTINGS_MODEL_MANAGE_BACKEND }}
+        </button>
       </div>
     </div>
-  </div>
-  <div class="border-b border-color-spilter flex flex-col gap-5 py-4">
-    <h2 class="text-center font-bold">Answer</h2>
-    <div class="flex flex-col gap-2">
-      <div class="flex items-center gap-2">
-        <div class="flex items-center gap-5">
-          <p>Performance Metrics</p>
-          <button
-            class="v-checkbox-control flex-none w-5 h-5"
-            :class="{ 'v-checkbox-checked': textInference.metricsEnabled }"
-            @click="textInference.toggleMetrics()"
-          ></button>
-        </div>
-      </div>
-      <div class="flex flex-col gap-2">
-        <p>Max Tokens</p>
-        <slide-bar
-          v-model:current="textInference.maxTokens"
-          :min="0"
-          :max="4096"
-          :step="1"
-        ></slide-bar>
-      </div>
+    <div class="text-right my-5">
+      <button @click="openDebug" class="v-radio-block">{{ languages.COM_DEBUG }}</button>
     </div>
-  </div>
-  <div class="flex flex-col gap-3">
-    <p>{{ languages.SETTINGS_BACKEND_STATUS }}</p>
-    <table class="text-center w-full mx-2 table-fixed">
-      <tbody>
-        <tr v-for="item in displayComponents" :key="item.serviceName">
-          <td style="text-align: left">{{ mapServiceNameToDisplayName(item.serviceName) }}</td>
-          <td :style="{ color: mapStatusToColor(item.status) }">
-            {{ mapToDisplayStatus(item.status) }}
-          </td>
-        </tr>
-      </tbody>
-    </table>
-    <div class="flex flex-col pt-5">
-      <button @click="globalSetup.loadingState = 'manageInstallations'" class="confirm-btn">
-        {{ languages.SETTINGS_MODEL_MANAGE_BACKEND }}
-      </button>
-    </div>
-  </div>
-  <div class="text-right my-5">
-    <button @click="openDebug" class="v-radio-block">{{ languages.COM_DEBUG }}</button>
   </div>
 </template>
 <script setup lang="ts">
@@ -145,22 +70,13 @@ import RadioBlock from '../components/RadioBlock.vue'
 import { useGlobalSetup } from '@/assets/js/store/globalSetup'
 
 import { useTheme } from '@/assets/js/store/theme'
-import { useTextInference, llmBackendTypes, LlmBackend } from '@/assets/js/store/textInference'
 import { mapServiceNameToDisplayName, mapStatusToColor, mapToDisplayStatus } from '@/lib/utils.ts'
 import { useBackendServices } from '@/assets/js/store/backendServices.ts'
 import LanguageSelector from '@/components/LanguageSelector.vue'
-import SlideBar from '../components/SlideBar.vue'
 
 const globalSetup = useGlobalSetup()
-const textInference = useTextInference()
 const backendServices = useBackendServices()
 const theme = useTheme()
-
-const textInferenceBackendDisplayName: Record<(typeof llmBackendTypes)[number], string> = {
-  ipexLLM: 'IPEX-LLM',
-  llamaCPP: 'llamaCPP - GGUF',
-  openVINO: 'OpenVINO',
-}
 
 const themeToDisplayName = (theme: Theme) => {
   switch (theme) {
@@ -197,22 +113,5 @@ function openDebug() {
 
 function changeGraphics(value: GraphicsItem, _index: number) {
   globalSetup.applyModelSettings({ graphics: value.index })
-}
-
-function mapBackendNames(name: LlmBackend): BackendServiceName | undefined {
-  if (name === 'ipexLLM') {
-    return 'ai-backend' as BackendServiceName
-  } else if (name === 'llamaCPP') {
-    return 'llamacpp-backend' as BackendServiceName
-  } else if (name === 'openVINO') {
-    return 'openvino-backend' as BackendServiceName
-  } else {
-    return undefined
-  }
-}
-
-function isRunning(name: LlmBackend) {
-  const backendName: BackendServiceName | undefined = mapBackendNames(name)
-  return backendServices.info.find((item) => item.serviceName === backendName)?.status === 'running'
 }
 </script>
